An attempt.
I don't expect that it compiles yet; --- I haven't tried. My chief
difficulty at the moment is that DragAndDrop.track requires a
Signal of whether the mouse is over the Ball, which in turn depends
on a Signal of the Ball. But I only create the Signal of the Ball
by folding over the signal returned by DragAndDrop.track! I may
have to switch to using the Automaton-based API.

The other thing I want to do is switch from Html to Graphics.
That's the only sane way to do it.

Onward!