Rewrote everything. I swear it almost compiles!
The reason that Git didn't catch the renaming of Ball.elm and
Wheeeeeeeeeeeeeeee.elm was that both source/Ball.elm and
source/Wheeeeeeeeeeeeeeee.elm bear only nominal resemblance to
their nominal predecessors. In particular, I've wormed my way out
of the Signal cycle by the magic of Mailboxes, which are basically
pipes: they're just a record of an Address (the write end) and a
Signal (the read end). Handy, no?

I've also switched from the cumbersome Html to the lighter and more
functional Graphics.*. A laudable decision, you say. I nod
politely; my mouth turns grim------But here arises a problem: The function that adds the
tracking of whether the cursor is over the ball operates on not a
Form but an Element. This is no good: Elements are rectangular, but
my ball must be a circle! Sure, it'll still *look* like a circle;
but when it comes to draggability it must *behave* as a circle! I
expect I'll have to wrangle my own equivalent to
Graphics.Input.hoverable. I may even resurrect that Euclidean
distance function I wrote for the last commit. I wouldn't mind
that. I was fond of that function. But we'll see.