Success!
Well, almost. You can drag the ball around, but not by clicking *on
the ball* and dragging: There's rather an offset, such that
dragging the cursor from the top left drags the ball which is in
the centre. Not to mention the y-movement is reversed --- but
that's an easy fix.