Switched Form to Ball; streamlined the code a little.
For the sake of doing away with this wonky and hacky business of
fiddling with library source files and poking around in internal
data constructors and deducing APIs from implementation details and
using sketchy case expressions etc., --- I have created a Ball
record, which knows its own radius and colour and position. When
the ball on the screen is dragged, instead of moving the Form, I
move the Ball, and transform the Ball into a Form in the view
function.  Slightly less efficient? Well, maybe, who knows. That is
in any case a lesser concern at this point. I suppose this is
probably more how Forms are intended to be used anyhow. Had I
thought this way earlier, I might not have wrote that rant about
types. ---Which in fact ended up having very little to do with my
problem with Forms at all; --- my ideas regarding type and data and
typeclasses and objects don't really apply here, --- though they do
apply elsewhere.

I also added a little bit of text in the middle of the window that
tells you whether the Elm program thinks the cursor is over the
ball or not, --- just for debugging. Wheeeeeeeeeeeeeeee.elm won't
run in the elm-reactor debugger ("type error: node.kids is
undefined"), probably because the debugger doesn't yet support
ports.

The call to Ball.relativeTo in the body of sender is commented out,
which I guess I didn't need to do. I was just trying to isolate the
bug. The bug being that you can only drag the ball once: drag it,
release it, and you can't pick it up again. The little bit of
debugging code described above tells me that when you drag the
ball, the hoverpad (draggable area --- should be simply the surface
of the ball, but it's still offset by half the window size, because
relativeTo is commented out) doesn't move along with it.
Mysterious. Maybe I'll try it with an Element, use
Graphics.Input.hoverable, see if that's any better. Otherwise I'm
baffled.

And I removed some unused imports from Ball.elm, and refactored
main a little. I guess that's about it.