Now you can replay macros too!
Drag the ball around while holding down 'r' to record a macro; then
press 'p' to replay it, starting from the ball's current position.
(It doesn't display the macro as text anymore. That was just for
testing.) The replay's not in real time, though; the ball just
jumps. But for most uses of macros that's usually what you want.

The problem is that every time you play back a macro, it consumes
an absurd amount of memory. You can drag the ball all you like,
you can record all the macros you please, and the ball still
faithfully keeps abreast of the cursor; but merely hit 'p' once or
twice and thereafter the ball's a stuttering zombie. I don't yet
know whether this is a problem with Wheeeeeee.elm or Macro.elm, ---
but whatever its origin and however tortuous its mechanism, I will
not rest until I bring this bug to rights!

---But, --- eager though I am to commence the hunt, that matter
shall wait for tomorrow, after sleep has replenished my faculties.
In the meantime, I'm just glad this works at all.