Macros can now be replayed.
The replay function takes a Macro and an Address and returns a Task
which sends each of the actions recorded in the macro to the
address in the order in which they were recorded. There might be a
bug that makes this function swallow absurd amounts of memory ---
or the bug might be here but somewhere else. I don't know yet.

I also made some of the code more readable and prettier.