Updated README.md regarding macros; bumped version.
Why bump the minor version? I don't know. It just felt right.

I hear evancz intends to have the elm compiler bump the version
number automatically, according to the extent of the changes to the
package's interface. It is a sensible scheme, and it'll mean a
decision like this would be not so much easier as not a decision at
all! I commend it.