Fixed a couple of longstanding bugs.
The first and most egregious bug was that the Signal that `record`
(and, by extension, `recordNamed`) returned was not a Signal of the
most recently recorded macro, as advertised, but rather a Signal of
the concatenation of all recorded macros. The problem itself was a
mere oversight: the `foldp` at the heart of `record` was not
designed to reset each time `record` was done recording, as it
should have been. See the notes in the Macro.elm for details.

The other bug was that the first action in a recorded macro was
always the most recent value of the source Signal *before*
recording began. Again, see the notes for details.