Split off point-related stuff into its own module.
Some of what's now in the new Point.elm  used to be under #
Coordinate stuff # in Wheeeeeee.elm. I figured (a) it has nothing
intrinsically to do with Wheeeeeee per se, rather it's all
essentially library functions; and (b) it became clear that
Ball.elm would need those and other functions that work with points
as well. Really this commit is just bookkeeping (and the slightest
bit of refactoring in Ball.elm --- I just didn't know when else I'd
be better off committing it).