Renamed functions in Tuple.elm.
To be precise, I renamed `mapBoth'` to `mapEach` and renamed
`mapBoth'` to `mapBoth`. I think the new names are a little more
evocative of the actual difference between the two functions and a
bit easier to remember. The convention of naming a variant by
appending a prime strikes me as simplistic and unhelpful. It is
true that naming things is one of the two hardest parts of
programming; but I refuse to take the easy way out. I have too much
pride for that.

Also fixed a naming inconsistency (Wheeeeeee.elm had `Ball.update`,
Ball.elm had `Ball.drag`) that I introduced in the last commit.