The cursor is now connected to the ball by an elastic band.
Mere commits ago, it seemed this project had pretty well fulfilled
its purpose. You could drag the ball around; and, what's more, you
could record the act of dragging and replay it at will. What joy!
what delight! For what more could one ask?

Then I thought: But what is that to me? This project is not and has
never been guided by what one might ask of it. No one has asked
anything, nor do I rely on someone's expectations; I have begotten
my schemes of my own fancy. Schemes, not expectations: I do not
demand, but experiment. I do not set goals and work to reach them,
but compose a design, and, in composing it, into it inject life.

This is all to say that connecting the cursor to the ball not
directly, as before, but by an elastic band, naturally arises as
the design which will inject the most life into the thing. And
lively it is! ---indeed, at the moment rather *too* lively. I
haven't yet added friction, is the thing; so the ball never loses
its momentum:---it just keeps swinging around. In fact, I've seen
it in this state actually *gain* momentum (perhaps due to the
vagaries of floating-point arithmetic).

But I will buff up these bristly bits in course. In the meantime,
my friends, do not hesitate to amuse yourselves with my little
bagatelle. You'd be a fool to ask more of it than that.