Added friction.
Now the ball is far less likely to zip out of control.

However, it does still zip out of control -- that is, continue to
accelerate, more than the friction can compensate for, -- if you
whip it too quickly, -- that is, if the acceleration's too great in
the first place. I think this may be because my simulation of the
mechanics of the elastic is too simplified.