Added dependency on elm-macro to elm-package.json.
Enough with this ad-hoc stuff---Enough with bypassing the package
manager---Enough with untracked versions and morphing
symlinks------Enough, I say! --But you'd think there'd be a way to
get elm-package to install a package from an arbitrary directory.
As it stands, apparently it can only fetch from the
package.elm-lang.org package catalogue. Oh well---it would be easy
enough to write a script for that, I guess.