Timed macros!
Macro.Timed has functions and types all analogous to those in
Macro,---the difference being that Macro.Timed deals with timed
macros --- where each action in the sequence is paired with a
timestamp, which represents the time elapsed at the actions's
recording since the previous actions was recorded --- in other
words, delta timestamps. Timed playback heeds these deltas --- that
is, a timed macro is played back with the timing with which it was
recorded.

So that's cool. I haven't tested it yet, though!