More documentation! Also factored out `and`.
Indeed, the documentation for the modules herein now takes the form
imposed by the standard which all Elm library modules are held up
to. Headings, examples, the works. I admire that compiler:---it
doesn't quaver in the face of the programmer's impatience,---no, it
stands by its principles!

And having that little `and` shoved in at the bottom of
Macro.elm was just silly. It now lives in Task.And,---which I've
half a mind to publish in its own package. I'd fork Task.Extra and
add it to that, but------why? `and` would no way contribute to
Task.Extra in particular, any more than it contributes to working
with Tasks in general. Indeed, it's precisely its absurd
succintness and its universal usefulness that make a package a poor
fit. A package would look rather ungainly and baggy on it; perhaps
a Gist would suit better. But, again because of `and`'s
nigh-unconstrained applicability, the convenience and integration
of the package manager wins out. ---------I'll think about it.