Bugfix! Replaying a macro no longer takes so dang long.
The bug: When a macro was replayed, there was a little delay before
the replay actually began. Morever, this delay grew longer with
each successively recorded macro. This is now fixed: The first
action in each macro is recorded with time-delta zero, which means
the time between record-begin and the first action is ignored; and
the time-deltas of successive actions are correct relative to the
first one.