Macros now preserve timing!
When you record and replay a macro, the relative timing of the
original actions is maintained. So if you hold down 'r' and drag
the ball around slow and then fast and then slow and then super
fast and then slow again, then when you hit 'p' the ball will move
slow and then fast and then slow and then super fast and then slow
again!

Also added explicit legit dependency on elm-macro; and removed a
couple unused imports.