I believe this is built from the 'svg' branch.
It feels a bit snappier than I remember the Form-based version
feeling. Maybe *that*'s why I decided to switch to svg in the first
place! Sorry it's black though; I haven't yet gotten around to
figuring out how to set the colour.

Well obviously this is more efficient: Svg is based on the virtual
dom.