Almost fix a bug in how friction is applied.
This is an inaccuracy in our simulation of friction. Currently the
ball slows down but never gets to the point of standing
still---it's always at least oscillating very slightly. Part of the
issue was we were slowing the ball down by the same amount even
when this was greater than the ball's velocity, in that case
actually speeding it up in the other direction. That's fixed, but
the oscillation is still happening. Probably this is because of the
way we compute the friction doesn't take into account the
acceleration due to the elastic.